:bento: 表禁用删除功能优化.

+++ hbase-manager-admin/src/main/java/com/leo/hbase/manager/web/controller/system/SysHbaseTableController.java
@@ -1,28 +1,27 @@
 package com.leo.hbase.manager.web.controller.system;
 
-import java.util.List;
-
+import com.leo.hbase.manager.adaptor.service.IHBaseAdminService;
+import com.leo.hbase.manager.common.annotation.Log;
+import com.leo.hbase.manager.common.core.controller.BaseController;
+import com.leo.hbase.manager.common.core.domain.AjaxResult;
+import com.leo.hbase.manager.common.core.page.TableDataInfo;
+import com.leo.hbase.manager.common.enums.BusinessType;
+import com.leo.hbase.manager.common.enums.HBaseDisabledFlag;
+import com.leo.hbase.manager.common.utils.poi.ExcelUtil;
+import com.leo.hbase.manager.framework.util.ShiroUtils;
+import com.leo.hbase.manager.system.domain.SysHbaseTable;
 import com.leo.hbase.manager.system.dto.SysHbaseTableDto;
 import com.leo.hbase.manager.system.service.ISysHbaseNamespaceService;
+import com.leo.hbase.manager.system.service.ISysHbaseTableService;
 import com.leo.hbase.manager.system.service.ISysHbaseTagService;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import com.leo.hbase.manager.common.annotation.Log;
-import com.leo.hbase.manager.common.enums.BusinessType;
-import com.leo.hbase.manager.system.domain.SysHbaseTable;
-import com.leo.hbase.manager.system.service.ISysHbaseTableService;
-import com.leo.hbase.manager.common.core.controller.BaseController;
-import com.leo.hbase.manager.common.core.domain.AjaxResult;
-import com.leo.hbase.manager.common.utils.poi.ExcelUtil;
-import com.leo.hbase.manager.common.core.page.TableDataInfo;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 /**
  * HBaseController
@@ -41,6 +40,8 @@ public class SysHbaseTableController extends BaseController {
     private ISysHbaseNamespaceService sysHbaseNamespaceService;
     @Autowired
     private ISysHbaseTagService sysHbaseTagService;
+    @Autowired
+    private IHBaseAdminService ihBaseAdminService;
 
     @RequiresPermissions("system:table:view")
     @GetMapping()
@@ -92,8 +93,9 @@ public class SysHbaseTableController extends BaseController {
     @PostMapping("/add")
     @ResponseBody
     public AjaxResult addSave(@Validated SysHbaseTableDto sysHbaseTableDto) {
-        //TODO 保存表信息时，同名表进行判断
         SysHbaseTable sysHbaseTable = sysHbaseTableDto.convertTo();
+
+        sysHbaseTable.setCreateBy(ShiroUtils.getSysUser().getLoginName());
         return toAjax(sysHbaseTableService.insertSysHbaseTable(sysHbaseTable));
     }
 
@@ -117,6 +119,15 @@ public class SysHbaseTableController extends BaseController {
     @PostMapping("/edit")
     @ResponseBody
     public AjaxResult editSave(SysHbaseTable sysHbaseTable) {
+        if (sysHbaseTable == null || sysHbaseTable.getTableId() < 1) {
+            return error("待修改信息的HBase表的id不能为空！");
+        }
+        SysHbaseTable exitsTable = sysHbaseTableService.selectSysHbaseTableById(sysHbaseTable.getTableId());
+        if (exitsTable == null || exitsTable.getTableId() < 1) {
+            return error("待修改信息的表[" + sysHbaseTable.getTableId() + "]不存在！");
+        }
+        sysHbaseTable.setUpdateBy(ShiroUtils.getSysUser().getLoginName());
+
         return toAjax(sysHbaseTableService.updateSysHbaseTable(sysHbaseTable));
     }
 
@@ -128,7 +139,28 @@ public class SysHbaseTableController extends BaseController {
     @PostMapping("/changeDisableStatus")
     @ResponseBody
     public AjaxResult changeDisableStatus(SysHbaseTable sysHbaseTable) {
-        return toAjax(sysHbaseTableService.updateSysHbaseTable(sysHbaseTable));
+        if (sysHbaseTable == null || sysHbaseTable.getTableId() < 1) {
+            return error("待修改表的id不能为空！");
+        }
+        SysHbaseTable exitsTable = sysHbaseTableService.selectSysHbaseTableById(sysHbaseTable.getTableId());
+        if (exitsTable == null || exitsTable.getTableId() < 1) {
+            return error("待修改的表[" + sysHbaseTable.getTableId() + "]不存在！");
+        }
+        final String fullHBaseTableName = exitsTable.getSysHbaseNamespace().getNamespaceName() + ":" + exitsTable.getTableName();
+        boolean changeTableDisabledStatusRes;
+        if (HBaseDisabledFlag.ENABLED.getCode().equals(sysHbaseTable.getDisableFlag())) {
+            changeTableDisabledStatusRes = ihBaseAdminService.enableTable(fullHBaseTableName);
+        } else if (HBaseDisabledFlag.DISABLED.getCode().equals(sysHbaseTable.getDisableFlag())) {
+            changeTableDisabledStatusRes = ihBaseAdminService.disableTable(fullHBaseTableName);
+        } else {
+            return error("暂时不支持的表禁用状态修改操作[" + sysHbaseTable.getDisableFlag() + "]");
+        }
+        if (!changeTableDisabledStatusRes) {
+            return error("系统异常，表状态修改失败！");
+        }
+
+        sysHbaseTable.setUpdateBy(ShiroUtils.getSysUser().getLoginName());
+        return toAjax(sysHbaseTableService.updateSysHbaseTableDisabledStatus(sysHbaseTable));
     }
 
 
@@ -140,8 +172,25 @@ public class SysHbaseTableController extends BaseController {
     @PostMapping("/remove")
     @ResponseBody
     public AjaxResult remove(Long ids) {
-        SysHbaseTable hbaseTable = sysHbaseTableService.selectSysHbaseTableById(ids);
-
+        SysHbaseTable sysHbaseTable = sysHbaseTableService.selectSysHbaseTableById(ids);
+        if (sysHbaseTable == null || sysHbaseTable.getTableId() < 1) {
+            return error("待删除表的id不能为空！");
+        }
+        SysHbaseTable exitsTable = sysHbaseTableService.selectSysHbaseTableById(sysHbaseTable.getTableId());
+        if (exitsTable == null || exitsTable.getTableId() < 1) {
+            return error("待删除的表[" + sysHbaseTable.getTableId() + "]不存在！");
+        }
+        if (HBaseDisabledFlag.ENABLED.getCode().equals(sysHbaseTable.getDisableFlag())) {
+            return error("非禁用状态的表不能被删除");
+        }
+        final String fullHBaseTableName = exitsTable.getSysHbaseNamespace().getNamespaceName() + ":" + exitsTable.getTableName();
+        if (!ihBaseAdminService.tableIsDisabled(fullHBaseTableName)) {
+            return error("非禁用状态的表不能被删除");
+        }
+        boolean deleteTableDisabledStatusRes = ihBaseAdminService.deleteTable(fullHBaseTableName);
+        if (!deleteTableDisabledStatusRes) {
+            return error("系统异常，表状态修改失败！");
+        }
         return toAjax(sysHbaseTableService.deleteSysHbaseTableById(ids));
     }
 }
